requirements.txt
Flask==2.2.3
Flask_Mail==0.9.1
psycopg2_binary==2.9.6
Werkzeug==2.2.3
